Make cleanup & terminate idempotent in KuberServantAllocator

+++ lzy-server/src/main/java/ru/yandex/cloud/ml/platform/lzy/server/kuber/allocators/KuberServantsAllocator.java
@@ -79,33 +79,43 @@ public class KuberServantsAllocator extends ServantsAllocatorBase {
     @Override
     protected void cleanup(ServantConnection s) {
         V1Pod pod = servantPods.get(s.id());
-        String name = Objects.requireNonNull(pod.getMetadata()).getName();
-        String namespace = pod.getMetadata().getNamespace();
-        try {
-            if (!isPodExists(namespace, name)) {
-                servantPods.remove(s.id());
-                return;
+        if (pod != null) {
+            String name = Objects.requireNonNull(pod.getMetadata()).getName();
+            String namespace = pod.getMetadata().getNamespace();
+            LOG.info("Cleaning up pod {} with connection {}", name, s.id());
+            try {
+                if (!isPodExists(namespace, name)) {
+                    servantPods.remove(s.id());
+                    return;
+                }
+            } catch (ApiException e) {
+                throw new RuntimeException(e);
             }
-        } catch (ApiException e) {
-            throw new RuntimeException(e);
+            terminate(s);
+        } else {
+            LOG.warn("Trying to cleanup nonexistent connection {}", s.id());
         }
-        terminate(s);
     }
 
     @Override
     protected void terminate(ServantConnection connection) {
         V1Pod pod = servantPods.get(connection.id());
-        String name = Objects.requireNonNull(pod.getMetadata()).getName();
-        String namespace = pod.getMetadata().getNamespace();
-        try {
-            if (isPodExists(namespace, name)) {
-                api.deleteNamespacedPod(name, namespace, null, null, 0,
-                        null, null, null);
+        if (pod != null) {
+            String name = Objects.requireNonNull(pod.getMetadata()).getName();
+            String namespace = pod.getMetadata().getNamespace();
+            LOG.info("Terminating up pod {} with connection {}", name, connection.id());
+            try {
+                if (isPodExists(namespace, name)) {
+                    api.deleteNamespacedPod(name, namespace, null, null, 0,
+                            null, null, null);
+                }
+            } catch (ApiException e) {
+                throw new RuntimeException(e);
             }
-        } catch (ApiException e) {
-            throw new RuntimeException(e);
+            servantPods.remove(connection.id());
+        } else {
+            LOG.warn("Trying to terminate nonexistent connection {}", connection.id());
         }
-        servantPods.remove(connection.id());
     }
 
     private boolean isPodExists(String namespace, String name) throws ApiException {
